vary dns-watcher's backend names
now that we are trying to avoid haproxy restarts by managing
the backends from the stats socket, we want to generate unique
backends names for different address values. we should assume
that if the name of the backend is the same, the address is also
the same.

this makes that change for the dns watcher, using the hash of the
backend address to generate the name.