centralize backend name construction
besides this being a more logical place for this function, it addresses a
potential bug where we were not using `construct_name` on default backends.
this means that if a default backend changed addresses but not names, we might
not catch it.

this bug is irrelevant right now, since a change in a default backend's address
involves reconfiguring and restarting synapse and in those cases we always
restart haproxy as well. but if we ever changed that behavior, it would
definitely bite.