support per-service config files
If the main config dict has a "service_conf_dir" key, use that value as
the name of a directory in which to search for multiple foo.json files.
Each file is parsed in turn, and appended to the main configuration dict
as a service named after the basename of the file, stripped of the .json
extension.