Make AWS region, access key, and secret key settings optional.
The AWS SDK is smart enough to figure out the region, access key, and secret
key if the user has not specified them. It looks in for the AWS_REGION environment
variable itself. It also uses a credential provider that attempts to locate
creditials from 1) AWS.config, 2) environment variables, 3) the EC2 metadata
service. Using the built-in configuration functionality of the SDK is required
for instances that get credentials based on IAM roles and have no access keys
on the instance.