Update README.md with better installation directions
The previous installation directions are not great -- they are ambiguous
between whether synapse should be used as a library (and placed in the
depending app's Gemfile) vs being used as a standalone application (and
installed via `gem install`).

The best way I know to deploy a Ruby application is to use `gem install
--install-dir [path]` to install the application in that path. This
allows:

* the gem to be installed by any user, not just root
* the version of synapse's dependencies to be isolated from other system
  gems
* easier redistribution as a system package

The downside is a relative large disk footprint - 36 MB as of writing.
Many files can be excluded from a system package, like the cache/
directory.