Add `backend_order` option support to `haproxy`
Currently we are doing shuffling (random ordering) only because we wanted to load
balance with round-robin to be more even across all machines.

This option allows one to specify how backends are ordered
in the final backend stanza configuration. This make it possible to support
usages that require an identical backend stanza configuration across machines.