Create option for using nerve weights
To be backwards compatible, it requires handling the use-case where users implmeneted HAProxy weighting using the `server_options` and `haproxy_server_options`.  This allows for continued normal functionality with the weight change.

Added logic to remove the weight options from `server_options` and `haproxy_server_options` when `use_nerve_weight` is enabled.  This is to prevent producing an odd haproxy backend configuration with multiple weights.